# Source_code/RestAPI/Definition/abdarum-libraryVarna-1.0.1-oas3-resolved.yaml
...
openapi: "3.0.0"
info:
  title: "Library Varna"
  description: "Library API"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "kornelstefanczyk@wp.pl"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
  version: "1.0.1-oas3"
...
servers:
- url: "http://localhost:8080"
security:
- bearerAuth: []
...
paths:
  /user:
    post:
      tags:
      - "user"
      summary: "Create user"
      description: "This can only be done by the logged in user."
      operationId: "createUser"
      requestBody:
        description: "Created user object"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        "403":
          description: "Forbidden, current user has no permission to create new user\
            \ of selected type"
        default:
          description: "successful operation"
  /user/login:
    get:
      tags:
      - "user"
      summary: "Logs user into the system"
      operationId: "loginUser"
      parameters:
      - name: "username"
        in: "query"
        description: "The user name for login"
        required: true
        style: "form"
        explode: true
        schema:
          type: "string"
      - name: "password"
        in: "query"
        description: "The password for login in clear text"
        required: true
        style: "form"
        explode: true
        schema:
          type: "string"
      responses:
        "200":
          description: "successful operation"
          headers:
            X-Rate-Limit:
              description: "calls per hour allowed by the user"
              style: "simple"
              explode: false
              schema:
                type: "integer"
                format: "int32"
            X-Expires-After:
              description: "date in UTC when token expires"
              style: "simple"
              explode: false
              schema:
                type: "string"
                format: "date-time"
          content:
            application/json:
              schema:
                type: "string"
        "400":
          description: "Invalid username/password supplied"
...
  /user/{username}:
    get:
      tags:
      - "user"
      summary: "Get user by user name"
      operationId: "getUserByName"
      parameters:
      - name: "username"
        in: "path"
        description: "The name that needs to be fetched. Use user1 for testing."
        required: true
        style: "simple"
        explode: false
        schema:
          type: "string"
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: "Invalid username supplied"
        "404":
          description: "User not found"
    put:
      tags:
      - "user"
      summary: "Updated user"
      description: "This can only be done by the logged in user."
      operationId: "updateUser"
      parameters:
      - name: "username"
        in: "path"
        description: "name that need to be updated"
        required: true
        style: "simple"
        explode: false
        schema:
          type: "string"
      requestBody:
        description: "Updated user object"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        "400":
          description: "Invalid user supplied"
        "404":
          description: "User not found"
    delete:
      tags:
      - "user"
      summary: "Delete user"
      description: "This can only be done by the logged in user."
      operationId: "deleteUser"
      parameters:
      - name: "username"
        in: "path"
        description: "The name that needs to be deleted"
        required: true
        style: "simple"
        explode: false
        schema:
          type: "string"
      responses:
        "400":
          description: "Invalid username supplied"
        "404":
          description: "User not found"
  /user/{username}/stats:
    get:
      tags:
      - "user"
      summary: "Get user statistics by user name"
      operationId: "getStatsByUsername"
      parameters:
      - name: "username"
        in: "path"
        required: true
        style: "simple"
        explode: false
        schema:
          type: "string"
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStatus"
        "400":
          description: "Invalid username supplied"
        "404":
          description: "User not found"
...
components:
  schemas:
    User:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        userTypeId:
          type: "integer"
          format: "int64"
        firstName:
          type: "string"
        lastName:
          type: "string"
        email:
          type: "string"
        password:
          type: "string"
        phone:
          type: "string"
        birthdate:
          type: "integer"
          description: "Timestamp in UTC"
        registrated:
          type: "integer"
          description: "Timestamp in UTC"
        gender:
          type: "string"
        adress:
          type: "string"
        city:
          type: "string"
        status:
          type: "string"
          description: "Order Status"
          enum:
          - "active"
          - "suspended"
          - "inactive"
      xml:
        name: "User"
...
    UserStatus:
      type: "object"
      properties:
        numberOfAllBorrowedBooks:
          type: "integer"
        numberOfAllDamagedBooks:
          type: "integer"
        numberOfCurrentBorrowedBooks:
          type: "integer"
        numberOfDelayedBooks:
          type: "integer"
      xml:
        name: "UserStatus"
...
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"